requirements.txt
annotated-types==0.7.0
anyio==4.3.0
blobfile==3.0.0
cachetools==5.5.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
et-xmlfile==1.1.0
fairscale==0.4.13
filelock==3.15.4
fire==0.6.0
fsspec==2024.6.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.67.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.24.6
idna==3.7
Jinja2==3.1.4
llama_models==0.0.10
llama_toolchain==0.0.10
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openpyxl==3.1.3
packaging==24.1
pandas==2.2.2
pillow==11.1.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.2
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rsa==4.9
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
sympy==1.13.2
termcolor==2.4.0
tiktoken==0.7.0
tk==0.1.0
torch==2.4.1
tqdm==4.66.5
ttkthemes==3.2.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
